+++ lib/decorate/engine.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var helpers = require('../helpers.js');
+var helpers = require('./helpers');
 var utils = require('../utils/');
 
 /**
