+++ lib/decorate/helpers.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var utils = require('./utils');
+var utils = require('../utils');
 
 module.exports = function initHelpers(proto) {
   init(proto);
